A bunch of commits I should have split
Not sure how to explain all the changes here. Most of it was done on the ATAT module with some minor code clean up.

The idea was for the position module to derive the necessary data for ATAT from it's existing data rather than dedicating entire methods in the ATAT module for this purpose.  The ATAT module should only focus on consuming the data.