Some refactoring
Down the line at in the ATAT module, we need to know from which sequence each kmer originated from. Previously, this was achieved by searching for each kmer within the original sequence file. This is incredibly resource intensive and simply takes too long.

This commit lays the groundwork for an implementation where the description of each initial sequence is stored in a class-level attribute and the Variants in each position are associated with an index number corresponding to their origin in the original sequence file. This index can then be used at any point to derive the full description of the sequence from which a specific variant originated from.